good progress, still have some failing tests and need to update query definition adapters to generate gql fragments with new syntax
