+++ src/dataUtilities.ts
@@ -1,21 +1,3 @@
-import { isArray, isObject } from 'lodash';
-import { OBJECT_PROPERTY_SEPARATOR } from './dataTypes';
-
-import {
-  FilterCondition,
-  EStringFilterOperator,
-  FilterValue,
-  INode,
-  SortObject,
-  ValidSortForNode,
-  FilterOperator,
-  QueryRecordEntry,
-  RelationalQueryRecordEntry,
-  DATA_TYPES,
-  IData,
-  RelationalQueryRecord,
-} from './types';
-
 /**
  * Clones an object or array. Recurses into nested objects and arrays for deep clones.
  */
@@ -134,194 +116,4 @@ export function extend(opts: {
         opts.object[extensionProp] = extensionValue;
       }
     });
-}
-
-/**
- * Returns all the keys of the object in a dot(.) separate format.
- *
- * Example.
- *
- * ```
- * getFlattenedObjectKeys({
- *  user: {
- *    firstName: 'John',
- *    lastName: 'Doe',
- *    meetings: [1,2,3],
- *    company: {
- *      name: 'Acme'
- *    }
- *  }
- * })
- * ```
- *
- * will return
- * ```
- * ['user.firstName', 'user.lastName', 'user.meetings', 'user.company.name']
- * ```
- *
- * Note: This won't flatten any array values
- *
- * @param obj - Object to flatten
- */
-export function getFlattenedObjectKeys(obj: Record<string, any>) {
-  const valuesByKeyPath = [];
-  for (const prop in obj) {
-    if (!obj.hasOwnProperty(prop)) continue;
-
-    // Exclude array values to prevent returning keys with array indexes
-    if (isObject(obj[prop]) && !isArray(obj[prop]) && obj[prop] !== null) {
-      const flatObject: any = getFlattenedObjectKeys(obj[prop]);
-      for (let idx = 0; idx < flatObject.length; idx++) {
-        valuesByKeyPath.push(`${prop}.${flatObject[idx]}`);
-      }
-    } else {
-      valuesByKeyPath.push(prop);
-    }
-  }
-  return valuesByKeyPath;
-}
-
-/**
- * Returns flattened keys of the filter object
- *
- * ```
- * getFlattenedNodeFilterObject({
- *  settings: {
- *    time: {_lte: Date.now()},
- *    nested: {
- *      prop: {_contains: "text"}
- *    }
- *  },
- *  firstName: {_eq: 'John'}
- * })
- * ```
- *
- * Returns
- *
- * ```
- * {
- *  "settings.time": {_lte: Date.now()},
- *  "settings.nested.prop": {_contains: "text"},
- *  "firstName": {_eq: 'John'}
- * }
- * ```
- * @param filterObject : ;
- * @returns
- */
-export function getFlattenedNodeFilterObject(opts: {
-  queryRecordEntry: QueryRecordEntry | RelationalQueryRecordEntry;
-}) {
-  const result: Record<
-    string,
-    Partial<Record<FilterOperator, any>> & {
-      condition: FilterCondition;
-    }
-  > = {};
-
-  const filterObject = opts.queryRecordEntry.filter;
-
-  if (!filterObject) return result;
-
-  const queriedRelations = opts.queryRecordEntry.relational;
-  const nodeData = opts.queryRecordEntry.def.data;
-
-  for (const filteredProperty in filterObject) {
-    const filterValue = filterObject[filteredProperty] as FilterValue<string>;
-
-    const isObjectInNodeData =
-      nodeData[filteredProperty] &&
-      ((nodeData[filteredProperty] as IData).type === DATA_TYPES.object ||
-        (nodeData[filteredProperty] as IData).type === DATA_TYPES.maybeObject);
-    const isAQueriedRelationalProp = queriedRelations
-      ? queriedRelations[filteredProperty] != null
-      : false;
-
-    const filterIsTargettingNestedObjectOrRelationalData =
-      isObject(filterValue) && (isAQueriedRelationalProp || isObjectInNodeData);
-
-    if (
-      typeof filterValue == 'object' &&
-      filterValue !== null &&
-      filterIsTargettingNestedObjectOrRelationalData
-    ) {
-      const queryRecordEntry = {
-        ...opts.queryRecordEntry,
-        def: isObjectInNodeData
-          ? {
-              ...opts.queryRecordEntry.def,
-              data: nodeData[filteredProperty].boxedValue,
-            }
-          : (queriedRelations as RelationalQueryRecord)[filteredProperty].def,
-        properties: isObjectInNodeData
-          ? opts.queryRecordEntry.properties
-              .filter(prop => prop.startsWith(filteredProperty))
-              .map(prop => {
-                const [, ...remainingPath] = prop.split(
-                  OBJECT_PROPERTY_SEPARATOR
-                );
-                return remainingPath.join(OBJECT_PROPERTY_SEPARATOR);
-              })
-          : (queriedRelations as RelationalQueryRecord)[filteredProperty]
-              .properties,
-        filter: filterValue,
-      };
-
-      const flatObject = getFlattenedNodeFilterObject({
-        queryRecordEntry,
-      });
-      Object.keys(flatObject).forEach(key => {
-        result[filteredProperty + '.' + key] = flatObject[key];
-      });
-    } else {
-      if (isObject(filterValue)) {
-        result[filteredProperty] = {
-          ...filterValue,
-          condition: filterValue.condition || 'and',
-        };
-      } else if (filterValue !== undefined) {
-        result[filteredProperty] = {
-          [EStringFilterOperator.eq]: filterValue,
-          condition: 'and',
-        } as Partial<Record<FilterOperator, any>> & {
-          condition: FilterCondition;
-        };
-      }
-    }
-  }
-  return result;
-}
-
-export function getFlattenedNodeSortObject<TNode extends INode>(
-  sorting: ValidSortForNode<TNode>
-) {
-  const result: Record<string, SortObject> = {};
-
-  for (const i in sorting) {
-    const sortObject = sorting as Record<string, any>;
-    const value = sortObject[i];
-    const valueIsNotASortObject =
-      isObject(value) && !Object.keys(value).includes('direction');
-    if (
-      typeof sortObject[i] == 'object' &&
-      sortObject[i] !== null &&
-      valueIsNotASortObject
-    ) {
-      const flatObject = getFlattenedNodeSortObject(value);
-      for (const x in flatObject) {
-        if (!flatObject.hasOwnProperty(x)) continue;
-
-        result[i + '.' + x] = flatObject[x];
-      }
-    } else {
-      if (isObject(value)) {
-        result[i] = value as SortObject;
-      } else if (value !== undefined) {
-        const filter: SortObject = {
-          direction: value,
-        };
-        result[i] = filter;
-      }
-    }
-  }
-  return result;
 }
