+++ src/queriers/clientSideOperators.ts
@@ -1,11 +1,7 @@
-import { update, isArray, orderBy } from 'lodash';
+import { update, isArray, isObject, orderBy } from 'lodash';
 import { NODES_PROPERTY_KEY } from '../consts';
-import { NULL_TAG } from '../dataConversions';
-import { OBJECT_PROPERTY_SEPARATOR } from '../dataTypes';
-import {
-  getFlattenedNodeFilterObject,
-  getFlattenedNodeSortObject,
-} from '../dataUtilities';
+import { OBJECT_PROPERTY_SEPARATOR } from '../queriers/queryDefinitionAdapters';
+
 import {
   FilterPropertyNotDefinedInQueryException,
   FilterOperatorNotImplementedException,
@@ -24,6 +20,10 @@ import {
   FilterCondition,
   RelationalQueryRecord,
   RelationalQueryRecordEntry,
+  IData,
+  DATA_TYPES,
+  EStringFilterOperator,
+  SortObject,
 } from '../types';
 
 function checkFilter({
@@ -95,18 +95,6 @@ function checkFilter({
   }
 }
 
-function convertNullStringValuesToNull({
-  item,
-  underscoreSeparatedPropName,
-}: {
-  item: any;
-  underscoreSeparatedPropName: string;
-}) {
-  return item[underscoreSeparatedPropName] === NULL_TAG
-    ? null
-    : item[underscoreSeparatedPropName];
-}
-
 function checkRelationalItems({
   relationalItems,
   operator,
@@ -119,7 +107,7 @@ function checkRelationalItems({
   filterValue: any;
 }) {
   return relationalItems.some(relationalItem => {
-    const relationalItemValue = convertNullStringValuesToNull({
+    const relationalItemValue = getValueWithUnderscoreSeparatedPropName({
       item: relationalItem,
       underscoreSeparatedPropName,
     });
@@ -224,7 +212,7 @@ export function applyClientSideFilterToData({
                 return filter.operators.every(({ operator, value }) => {
                   if (filter.oneToOne === true) {
                     const itemValue = filter.relationalKey
-                      ? convertNullStringValuesToNull({
+                      ? getValueWithUnderscoreSeparatedPropName({
                           item: item[filter.relationalKey],
                           underscoreSeparatedPropName:
                             filter.underscoreSeparatedPropName,
@@ -250,10 +238,7 @@ export function applyClientSideFilterToData({
                   }
                 });
               } else {
-                const itemValue =
-                  item[filter.underscoreSeparatedPropName] === NULL_TAG
-                    ? null
-                    : item[filter.underscoreSeparatedPropName];
+                const itemValue = item[filter.underscoreSeparatedPropName];
                 return filter.operators.every(({ operator, value }) => {
                   return checkFilter({
                     operator,
@@ -274,7 +259,7 @@ export function applyClientSideFilterToData({
                 return filter.operators.some(({ operator, value }) => {
                   if (filter.oneToOne === true) {
                     const itemValue = filter.relationalKey
-                      ? convertNullStringValuesToNull({
+                      ? getValueWithUnderscoreSeparatedPropName({
                           item: item[filter.relationalKey],
                           underscoreSeparatedPropName:
                             filter.underscoreSeparatedPropName,
@@ -301,7 +286,7 @@ export function applyClientSideFilterToData({
                 });
               } else {
                 const itemValue = filter.relationalKey
-                  ? convertNullStringValuesToNull({
+                  ? getValueWithUnderscoreSeparatedPropName({
                       item,
                       underscoreSeparatedPropName:
                         filter.underscoreSeparatedPropName,
@@ -381,14 +366,13 @@ function getNodeSortPropertyValue(opts: {
 }
 
 function getItemSortValue(item: any, underscoreSeparatedPropertyPath: string) {
-  const isValueNull =
-    item[underscoreSeparatedPropertyPath] === null ||
-    item[underscoreSeparatedPropertyPath] === NULL_TAG;
+  const value = getValueWithUnderscoreSeparatedPropName({
+    item,
+    underscoreSeparatedPropName: underscoreSeparatedPropertyPath,
+  });
+  const isValueNull = value === null;
   if (isValueNull) return null;
-  return (
-    Number(item[underscoreSeparatedPropertyPath]) ||
-    item[underscoreSeparatedPropertyPath]
-  );
+  return Number(value) || value;
 }
 
 export function applyClientSideSortToData({
@@ -538,4 +522,169 @@ export function applyClientSideSortAndFilterToData(
       }
     }
   });
+}
+
+/**
+ * Returns flattened keys of the filter object
+ *
+ * ```
+ * getFlattenedNodeFilterObject({
+ *  settings: {
+ *    time: {_lte: Date.now()},
+ *    nested: {
+ *      prop: {_contains: "text"}
+ *    }
+ *  },
+ *  firstName: {_eq: 'John'}
+ * })
+ * ```
+ *
+ * Returns
+ *
+ * ```
+ * {
+ *  "settings.time": {_lte: Date.now()},
+ *  "settings.nested.prop": {_contains: "text"},
+ *  "firstName": {_eq: 'John'}
+ * }
+ * ```
+ * @param filterObject : ;
+ * @returns
+ */
+function getFlattenedNodeFilterObject(opts: {
+  queryRecordEntry: QueryRecordEntry | RelationalQueryRecordEntry;
+}) {
+  const result: Record<
+    string,
+    Partial<Record<FilterOperator, any>> & {
+      condition: FilterCondition;
+    }
+  > = {};
+
+  const filterObject = opts.queryRecordEntry.filter;
+
+  if (!filterObject) return result;
+
+  const queriedRelations = opts.queryRecordEntry.relational;
+  const nodeData = opts.queryRecordEntry.def.data;
+
+  for (const filteredProperty in filterObject) {
+    const filterValue = filterObject[filteredProperty] as FilterValue<string>;
+
+    const isObjectInNodeData =
+      nodeData[filteredProperty] &&
+      ((nodeData[filteredProperty] as IData).type === DATA_TYPES.object ||
+        (nodeData[filteredProperty] as IData).type === DATA_TYPES.maybeObject);
+    const isAQueriedRelationalProp = queriedRelations
+      ? queriedRelations[filteredProperty] != null
+      : false;
+
+    const filterIsTargettingNestedObjectOrRelationalData =
+      isObject(filterValue) && (isAQueriedRelationalProp || isObjectInNodeData);
+
+    if (
+      typeof filterValue == 'object' &&
+      filterValue !== null &&
+      filterIsTargettingNestedObjectOrRelationalData
+    ) {
+      const queryRecordEntry = {
+        ...opts.queryRecordEntry,
+        def: isObjectInNodeData
+          ? {
+              ...opts.queryRecordEntry.def,
+              data: nodeData[filteredProperty].boxedValue,
+            }
+          : (queriedRelations as RelationalQueryRecord)[filteredProperty].def,
+        properties: isObjectInNodeData
+          ? opts.queryRecordEntry.properties
+              .filter(prop => prop.startsWith(filteredProperty))
+              .map(prop => {
+                const [, ...remainingPath] = prop.split(
+                  OBJECT_PROPERTY_SEPARATOR
+                );
+                return remainingPath.join(OBJECT_PROPERTY_SEPARATOR);
+              })
+          : (queriedRelations as RelationalQueryRecord)[filteredProperty]
+              .properties,
+        filter: filterValue,
+      };
+
+      const flatObject = getFlattenedNodeFilterObject({
+        queryRecordEntry,
+      });
+      Object.keys(flatObject).forEach(key => {
+        result[filteredProperty + '.' + key] = flatObject[key];
+      });
+    } else {
+      if (isObject(filterValue)) {
+        result[filteredProperty] = {
+          ...filterValue,
+          condition: filterValue.condition || 'and',
+        };
+      } else if (filterValue !== undefined) {
+        result[filteredProperty] = {
+          [EStringFilterOperator.eq]: filterValue,
+          condition: 'and',
+        } as Partial<Record<FilterOperator, any>> & {
+          condition: FilterCondition;
+        };
+      }
+    }
+  }
+  return result;
+}
+
+function getFlattenedNodeSortObject<TNode extends INode>(
+  sorting: ValidSortForNode<TNode>
+) {
+  const result: Record<string, SortObject> = {};
+
+  for (const i in sorting) {
+    const sortObject = sorting as Record<string, any>;
+    const value = sortObject[i];
+    const valueIsNotASortObject =
+      isObject(value) && !Object.keys(value).includes('direction');
+    if (
+      typeof sortObject[i] == 'object' &&
+      sortObject[i] !== null &&
+      valueIsNotASortObject
+    ) {
+      const flatObject = getFlattenedNodeSortObject(value);
+      for (const x in flatObject) {
+        if (!flatObject.hasOwnProperty(x)) continue;
+
+        result[i + '.' + x] = flatObject[x];
+      }
+    } else {
+      if (isObject(value)) {
+        result[i] = value as SortObject;
+      } else if (value !== undefined) {
+        const filter: SortObject = {
+          direction: value,
+        };
+        result[i] = filter;
+      }
+    }
+  }
+  return result;
+}
+
+function getValueWithUnderscoreSeparatedPropName(opts: {
+  item: any;
+  underscoreSeparatedPropName: string;
+}): any {
+  const { item, underscoreSeparatedPropName } = opts;
+  const [currentProperty, ...remainingPath] = underscoreSeparatedPropName.split(
+    OBJECT_PROPERTY_SEPARATOR
+  );
+  if (remainingPath.length === 0) {
+    return item[currentProperty];
+  } else {
+    return getValueWithUnderscoreSeparatedPropName({
+      item: item[currentProperty],
+      underscoreSeparatedPropName: remainingPath.join(
+        OBJECT_PROPERTY_SEPARATOR
+      ),
+    });
+  }
 }
