--- src/transaction/edges/index.ts
@@ -1,23 +1,0 @@
-export {
-  createEdge,
-  createEdges,
-  getMutationsFromEdgeCreateOperations,
-} from './create';
-
-export {
-  dropEdge,
-  dropEdges,
-  getMutationsFromEdgeDropOperations,
-} from './drop';
-
-export {
-  replaceEdge,
-  replaceEdges,
-  getMutationsFromEdgeReplaceOperations,
-} from './replace';
-
-export {
-  updateEdge,
-  updateEdges,
-  getMutationsFromEdgeUpdateOperations,
-} from './update';
