+++ src/types.ts
@@ -4,7 +4,6 @@ import { createDOFactory } from './DO';
 import { createDOProxyGenerator } from './DOProxyGenerator';
 import { generateQuerier, generateSubscriber } from './queriers';
 import { createQueryManager } from './queriers/QueryManager';
-import { createTransaction } from './transaction/transaction';
 import { QuerySlimmer } from './queriers/QuerySlimmer'
 
 export type BOmit<T, K extends keyof T> = T extends any ? Omit<T, K> : never;
@@ -167,7 +166,6 @@ export interface IMMGQL {
   clearTokens(): void
   query: ReturnType<typeof generateQuerier>
   subscribe: ReturnType<typeof generateSubscriber>
-  transaction: ReturnType<typeof createTransaction>
   gqlClient: IGQLClient
   plugins: Array<Plugin> | undefined
   generateMockData: boolean | undefined
